Move Notation algorithms into their own files
